requirements.txt
argon2-cffi
django-cors-headers
djangorestframework
pandas
